requirements.txt
ccxt
fastapi
httpx
line-bot-sdk
mplfinance
numpy
pandas
pandas-ta